refactor: remove redundant Python Session wrapper, use PyO3 generated Session directly
Similar to the TypeScript changes, the PyO3 generated Session class provides an excellent
Python API with proper methods, properties, and docstrings, making the wrapper redundant.

This simplifies the codebase by:
- Removing the Python Session wrapper class
- Importing Session directly from _lancedb module
- Using PyO3 generated Session with native constructor and methods
- Maintaining identical user-facing API with better performance

The PyO3 Session provides:
- Constructor: Session(index_cache_size_bytes=None, metadata_cache_size_bytes=None)
- Static method: Session.default()
- Properties: size_bytes, approx_num_items
- Full docstrings preserved from Rust implementation
